Refine seedDB script for Epsilon and upcoming classes
