+++ frontend-v2/src/util/curr-pledge-class.js
@@ -1,6 +1,6 @@
 import { PledgeClassEnum } from './enums/brother-enums';
 
 /* No pledge class: set to null, otherwise use the enum */
-const CURR_PLEDGE_CLASS = PledgeClassEnum.DELTA;
+const CURR_PLEDGE_CLASS = PledgeClassEnum.EPSILON;
 
 export default CURR_PLEDGE_CLASS;
