+++ server/routes/api/util/brother-info/officersF21.js
@@ -1,0 +1,203 @@
+const {
+  MajorEnum,
+  PledgeClassEnum,
+  PositionEnum,
+} = require('../enums/brother-enums');
+
+const officers = [
+  {
+    name: 'Kevin Lin',
+    email: 'kevinnlin28@gmail.com',
+    studentID: '',
+    phoneNumber: '6263897495',
+    major: MajorEnum.CIVIL,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.ALPHA,
+    position: PositionEnum.PRESIDENT,
+    isGraduated: false,
+  },
+  {
+    name: 'Ashana Patel',
+    email: 'ashana.patel@sjsu.edu',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.ELECTRICAL,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.VICE_PRESIDENT,
+    isGraduated: false,
+  },
+  {
+    name: 'Christopher Tom',
+    email: 'christophertom2000@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.ACADEMIC,
+    isGraduated: false,
+  },
+  {
+    name: 'Ian Crane',
+    email: 'ianmcrane1@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.ELECTRICAL,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.TREASURER,
+    isGraduated: false,
+  },
+  {
+    name: 'Derek Huang',
+    email: 'dchuang14@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.CIVIL,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.SECRETARY,
+    isGraduated: false,
+  },
+  {
+    name: 'Anthony Lal',
+    email: 'anthony.lal@sjsu.edu',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.CNS,
+    graduatingYear: 2021,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.RISK_MANAGEMENT,
+    isGraduated: false,
+  },
+  {
+    name: 'Sunny Chen',
+    email: 'sunny.m.chen@sjsu.edu',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.HISTORIAN,
+    isGraduated: false,
+  },
+  {
+    name: 'Kevin Chu',
+    email: 'kchu89@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.ALPHA,
+    position: PositionEnum.PLEDGE_PARENT,
+    isGraduated: false,
+  },
+  {
+    name: 'Parker Grube',
+    email: 'parkergrube@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.MECHANICAL,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.PLEDGE_PARENT,
+    isGraduated: false,
+  },
+  {
+    name: 'Evelyn Chan',
+    email: 'evelyn.chan@sjsu.edu',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.MECHANICAL,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.FUNDRAISING,
+    isGraduated: false,
+  },
+  {
+    name: 'Hitayshi Shah',
+    email: 'hitayshi27@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.SOFTWARE,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.COMMUNITY_SERVICE,
+    isGraduated: false,
+  },
+  {
+    name: 'Neel Thakkar',
+    email: 'neelt11@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.SOFTWARE,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.PROFESSIONALISM,
+    isGraduated: false,
+  },
+  {
+    name: 'Frank Ka',
+    email: 'frka0526@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.CIVIL,
+    graduatingYear: 2024,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.BROTHERHOOD,
+    isGraduated: false,
+  },
+  {
+    name: 'Eloisa Difuntorum',
+    email: 'eloisakayed@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.ELECTRICAL,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.GAMMA,
+    position: PositionEnum.RUSH,
+    isGraduated: false,
+  },
+  {
+    name: 'Nicholas Mendoza',
+    email: 'nikkomdoza@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.SOFTWARE,
+    graduatingYear: 2022,
+    pledgeClass: PledgeClassEnum.BETA,
+    position: PositionEnum.RUSH,
+    isGraduated: false,
+  },
+  {
+    name: 'Devonna Dang',
+    email: 'dang.devonna@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.PUBLIC_RELATIONS,
+    isGraduated: false,
+  },
+  {
+    name: 'Sahil Patel',
+    email: 'patelsahil395@gmail.com',
+    studentID: '',
+    phoneNumber: '',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.DELTA,
+    position: PositionEnum.WEBMASTER,
+    isGraduated: false,
+  },
+
+
+
+
+
+
+];
+
+module.exports = officers;
