+++ server/routes/api/util/brother-info/pledgesF21info.js
@@ -1,0 +1,169 @@
+const {
+  MajorEnum,
+  PledgeClassEnum,
+  PositionEnum,
+} = require('../enums/brother-enums');
+
+const pledgeBrother = [
+  {
+    name: 'Alberto Meunier',
+    email: 'biiirdeaux@gmail.com',
+    studentID: '015484638',
+    phoneNumber: '2027108129',
+    major: MajorEnum.AEROSPACE,
+    graduatingYear: 2024,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Akshin Gopal',
+    email: 'akshingopal@gmail.com',
+    studentID: '013885014',
+    phoneNumber: '4088196186',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Breanna Chi',
+    email: 'breannakchi@gmail.com',
+    studentID: '014384955',
+    phoneNumber: '9252166350',
+    major: MajorEnum.AEROSPACE,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Trevor Pedron',
+    email: 'pedrontrevor@gmail.com',
+    studentID: '013166179',
+    phoneNumber: '2096066205',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Novel Alam',
+    email: 'nonyzeon@gmail.com',
+    studentID: '015582788',
+    phoneNumber: '4085104387',
+    major: MajorEnum.CMPE,
+    graduatingYear: 2025,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Russell Semsem',
+    email: 'semsemrussell@gmail.com',
+    studentID: '015516683',
+    phoneNumber: '6692255859',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2025,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Karan Gandhi',
+    email: 'freekaran@gmail.com',
+    studentID: '014986816',
+    phoneNumber: '5107714975',
+    major: MajorEnum.SOFTWARE,
+    graduatingYear: 2024,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Julian Obusan',
+    email: 'julianobusan@gmail.com',
+    studentID: '014038856',
+    phoneNumber: '4083388338',
+    major: MajorEnum.CIVIL,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Quinmann Young',
+    email: 'qmyoung@att.net',
+    studentID: '013992953',
+    phoneNumber: '9258586280',
+    major: MajorEnum.AEROSPACE,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'James Yu',
+    email: 'jamesyu200141@gmail.com',
+    studentID: '014054599',
+    phoneNumber: '4087280491',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'William Xuan',
+    email: 'xuanwilliam8@gmail.com',
+    studentID: '015780323',
+    phoneNumber: '4086746195',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2025,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Anwar Mujeeb',
+    email: 'ibnmujeeb2003@gmail.com',
+    studentID: '015593851',
+    phoneNumber: '6692120331',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2025,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'John Vaughn',
+    email: 'mickvaughn56@gmail.com',
+    studentID: '013883688',
+    phoneNumber: '8432959665',
+    major: MajorEnum.BIOMEDICAL,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+  {
+    name: 'Tiffany Phan',
+    email: 'phanntiffany@gmail.com',
+    studentID: '014112969',
+    phoneNumber: '6692378688',
+    major: MajorEnum.COMP_SCI,
+    graduatingYear: 2023,
+    pledgeClass: PledgeClassEnum.EPSILON,
+    position: PositionEnum.PLEDGE,
+    isGraduated: false,
+  },
+
+
+
+
+
+];
+
+module.exports = pledgeBrother;
