+++ server/server.js
@@ -6,11 +6,8 @@ const fs = require('fs');
 const cors = require('cors');
 const AWS = require('aws-sdk');
 
-// const seedDB = require('./routes/api/util/scripts/seed-database');
-// const changeBrotherInformation = require('./routes/api/util/scripts/change-brother-info');
-const addEpsilon = require('./routes/api/util/scripts/add-epsilon');
-const removeDelta = require('./routes/api/util/scripts/reset-delta');
-const updatePos = require('./routes/api/util/scripts/update-pos');
+const seedDB = require('./routes/api/util/scripts/seed-database');
+
 // Allows requests from diff origins
 app.use(cors());
 
@@ -38,16 +35,15 @@ switch (process.env.NODE_ENV) {
         useNewUrlParser: true,
         useUnifiedTopology: true,
         useCreateIndex: true,
+        useFindAndModify: false,
       })
-      .then(() => console.log('Development mode: Connected to MongoDB Atlas'))
+      .then(() => {
+        console.log('DEVELOPMENT MODE: Connected to MongoDB Atlas\n');
+        seedDB();
+      })
       .catch((error) => {
         throw new Error(error);
       });
-    // seedDB();
-    // changeBrotherInformation();
-    removeDelta();
-    //addEpsilon();
-    updatePos();
     break;
   case 'production':
     mongoose
@@ -55,15 +51,15 @@ switch (process.env.NODE_ENV) {
         useNewUrlParser: true,
         useUnifiedTopology: true,
         useCreateIndex: true,
+        useFindAndModify: false,
       })
-      .then(() =>
-        console.log('Production mode: Connected to MongoDB container')
-      )
+      .then(() => {
+        console.log('PRODUCTION MODE: Connected to MongoDB container\n');
+        seedDB();
+      })
       .catch((error) => {
         throw new Error(error);
       });
-    // seedDB();
-    // changeBrotherInformation();
     break;
   default:
     console.error('NODE_ENV not specified');
